+++ base/usr/include/setjmp.h
@@ -7,7 +7,7 @@ _Begin_C_Header
 /* i386 */
 #define _JBLEN 9
 
-typedef int jmp_buf[_JBLEN];
+typedef long long jmp_buf[_JBLEN];
 
 extern void longjmp(jmp_buf j, int r);
 extern int setjmp(jmp_buf j);
