+++ libc/unistd/read.c
@@ -3,8 +3,8 @@
 #include <syscall.h>
 #include <syscall_nums.h>
 
-DEFN_SYSCALL3(read,  SYS_READ, int, char *, int);
+DEFN_SYSCALL3(read,  SYS_READ, int, char *, size_t);
 
-int read(int file, void *ptr, size_t len) {
+ssize_t read(int file, void *ptr, size_t len) {
 	__sets_errno(syscall_read(file,ptr,len));
 }
