+++ libc/unistd/write.c
@@ -3,7 +3,7 @@
 #include <syscall.h>
 #include <syscall_nums.h>
 
-DEFN_SYSCALL3(write, SYS_WRITE, int, char *, int);
+DEFN_SYSCALL3(write, SYS_WRITE, int, char *, size_t);
 
 ssize_t write(int file, const void *ptr, size_t len) {
 	__sets_errno(syscall_write(file,(char *)ptr,len));
